requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-gcp>=6.0.0,<7.0.0
google-cloud-containeranalysis==2.12.1
